Simplify error handling in TransFlickr by changing to a more pythonic error handling scheme in flickrapi.  Instead of a method we pass the result XMLNode back to for parsing, put the methods on the XMLNode itself.  Two new methods: __nonzero__, which returns True if the XMLNode contains a valid response, and False if it contains an error. That way we can say "if rsp: <good case> else: <error case>".  Next we add a get_errortext method that parses out the error message and returns it.  Then we assign that as a property to 'errormsg'.  So now our error case can say "log.error("error: %s" % rsp.errormsg)".
I don't have actual test cases for handling errors.  In its current
state the code generates one ERROR in the log, so this tests that
the new XMLNode code is working correctly.  The test suite tests that
the True case for __nonzero__ works in all the tested code.  The
error cases are a very straightforward replacement of "retinfo"
by "rsp.errormsg", so it is unlikely I introduced any errors there.

In fact I fixed one in getBandwidthInfo, since if that code had ever
returned an error it would have bombed since there was code that
referenced the contents of the XMLNode between the call to the
old returntestFailure and the if test of the retinfo.

I also changed a couple of my new TransFlickr routines that had been
using a name other than 'rsp' for the results of the fapi call to
use rsp.  The other names were left over from when the code was
living on the flickrfs object.  This way, the code is consistent
and easier to maintain.